+++ CHANGELOG.md
@@ -81,6 +81,7 @@
 ### Tools
 
 * [BUGFIX] Stop tools from panicking when `-help` flag is passed. #5412
+* [BUGFIX] Remove github.com/golang/glog command line flags from tools. #5413
 
 ## 2.9.0